+++ CrossRename/main.py
@@ -4,8 +4,10 @@ import re
 from pathlib import Path
 import argparse
 import logging
+from .utils import check_for_update
 
-__version__ = "1.0.0"
+__version__ = "1.1.0"
+
 
 logger = logging.getLogger(__name__)
 logging.basicConfig(level=logging.INFO, format='%(asctime)s - %(levelname)s » %(message)s')
@@ -61,26 +63,44 @@ def sanitize_filename(filename: str) -> str:
     return sanitized
 
 
-def rename_file(file_path: str) -> None:
+def rename_file(file_path: str, dry_run: bool = False) -> None:
     directory, filename = os.path.split(file_path)
     new_filename = sanitize_filename(filename)
 
     if new_filename != filename:
         new_file_path = os.path.join(directory, new_filename)
-        try:
-            os.rename(file_path, new_file_path)
-            logger.info(f"Renamed: {filename} -> {new_filename}")
-        except Exception as e:
-            logger.error(f"Error renaming {filename}: {str(e)}")
+        if dry_run:
+            logger.info(f"[Dry-run] Would rename: {filename} -> {new_filename}")
+        else:
+            try:
+                os.rename(file_path, new_file_path)
+                logger.info(f"Renamed: {filename} -> {new_filename}")
+            except Exception as e:
+                logger.error(f"Error renaming {filename}: {str(e)}")
     else:
         logger.info(f"No change needed: {filename}")
 
 
 def file_search(directory: str) -> list[str]:
     file_list = []
-    for root, _, files in os.walk(directory):
+    visited_paths = set()
+
+    for root, _, files in os.walk(directory, followlinks=False):
+        real_root = os.path.realpath(root)
+
+        if real_root in visited_paths:
+            logger.warning(f"Skipping recursive symlink in {root}")
+            continue
+
+        visited_paths.add(real_root)
+
         for file in files:
-            file_list.append(os.path.join(root, file))
+            file_path = os.path.join(root, file)
+            if os.path.islink(file_path):
+                logger.info(f"Skipping symlink: {file_path}")
+                continue
+            file_list.append(file_path)
+
     return file_list
 
 
@@ -88,7 +108,7 @@ def main() -> None:
     try:
         parser = argparse.ArgumentParser(
             description="CrossRename: Harmonize file names for Linux and Windows.")
-        parser.add_argument("-p", "--path", help="The path to the file or directory to rename.", required=True)
+        parser.add_argument("-p", "--path", help="The path to the file or directory to rename.")
         parser.add_argument(
             "-v",
             "--version",
@@ -97,31 +117,43 @@ def main() -> None:
             version=f"CrossRename Version {__version__}"
         )
         parser.add_argument(
+            "-u", "--update",
+            help="Check if a new version is available.",
+            action="store_true"
+        )
+        parser.add_argument(
             "-r",
             "--recursive",
             help="Rename all files in the directory path given and its subdirectories.",
             action="store_true"
         )
-        args = parser.parse_args()
+        parser.add_argument("-d", "--dry-run", help="Perform a dry run, logging changes without renaming.",
+                            action="store_true")
 
+        args = parser.parse_args()
         path = args.path
         recursive = args.recursive
+        dry_run = args.dry_run
+
+        if args.update:
+            check_for_update(__version__)
+            sys.exit()
 
         if path is None:
-            sys.exit("Please provide a path to a file or directory using the --path argument.")
+            sys.exit("Error: Please provide a path to a file or directory using the --path argument.")
 
         if os.path.isfile(path):
-            rename_file(path)
+            rename_file(path, dry_run)
         elif os.path.isdir(path):
             if recursive:
                 file_list = file_search(path)
                 for file_path in file_list:
-                    rename_file(file_path)
+                    rename_file(file_path, dry_run)
             else:
                 for item in os.listdir(path):
                     item_path = os.path.join(path, item)
                     if os.path.isfile(item_path):
-                        rename_file(item_path)
+                        rename_file(item_path, dry_run)
         else:
             sys.exit(f"Error: {path} is not a valid file or directory")
     except Exception as e: